Refactor: 업비트 함수 분리
