+++ src/utils/priceCalc.ts
@@ -1,5 +1,16 @@
+import { IUpbitTicker } from 'api/upbit/useUpbitTicker';
 import { IBithumbFetchTicker } from 'components/bithumb/Bithumb.type';
 
+// 업비트
+// 고가 대비 증감률(52주)
+export const highest_52_week_rate = (data: IUpbitTicker) =>
+  (data.trade_price / data.highest_52_week_price - 1) * 100;
+
+// 저가 대비 증감률(52주)
+export const lowest_52_week_rate = (data: IUpbitTicker) =>
+  (data.trade_price / data.lowest_52_week_price - 1) * 100;
+
+// 빗썸
 // 전일 대비 증감률
 export const changesRatio = (data: IBithumbFetchTicker) =>
   ((Number(data[1].closing_price) - Number(data[1].prev_closing_price)) /
